automation/requirements.txt
# LinkedIn Content Automation Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.31.0
pyyaml>=6.0
pathlib2>=2.3.7

# Content analysis and enhancement
openai>=1.0.0
anthropic>=0.25.0

# Web scraping and research
beautifulsoup4>=4.12.0
selenium>=4.15.0
newspaper3k>=0.2.8

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Scheduling and automation
schedule>=1.2.0
python-crontab>=2.7.0

# Image processing (for future visual content)
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API integrations
google-api-python-client>=2.100.0
tweepy>=4.14.0  # For future Twitter integration

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.5.0  # For beautiful console output
tqdm>=4.66.0  # Progress bars

# Development and testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Optional: For advanced content analysis
# nltk>=3.8.1
# spacy>=3.6.0
# transformers>=4.30.0
